- merge in support for secondary hubs for Pi 2/3 - add support for secondary hubs in Pi 3+

+++ scripts/rmlinkandtarget.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 # remove the target of a symlink *and* the link itself
-
-rm -f "`readlink $1`"
-rm -f "$1"
+# if they don't exist, set
+if [[ -L "$1" ]]; then
+    rm -f "`readlink $1`"
+    rm -f "$1"
+    exit 0;
+else
+    exit 1;
+fi
